sha1: refactor folder structure
Eliminate the `sha1` folder, move it down into `hash` so that future
impelmentations can share common code.
