+++ src/util/hash/collisiondetect.c
