+++ src/util/hash/collisiondetect.h
@@ -5,15 +5,15 @@
  * a Linking Exception. For full terms see the included COPYING file.
  */
 
-#ifndef INCLUDE_hash_sha1_common_crypto_h__
-#define INCLUDE_hash_sha1_common_crypto_h__
+#ifndef INCLUDE_hash_collisiondetect_h__
+#define INCLUDE_hash_collisiondetect_h__
 
-#include "hash/sha1.h"
+#include "hash/sha.h"
 
-#include <CommonCrypto/CommonDigest.h>
+#include "sha1dc/sha1.h"
 
 struct git_hash_sha1_ctx {
-	CC_SHA1_CTX c;
+	SHA1_CTX c;
 };
 
 #endif
