+++ src/util/hash/common_crypto.c
