+++ src/util/hash/common_crypto.h
@@ -5,15 +5,15 @@
  * a Linking Exception. For full terms see the included COPYING file.
  */
 
-#ifndef INCLUDE_hash_sha1_mbedtls_h__
-#define INCLUDE_hash_sha1_mbedtls_h__
+#ifndef INCLUDE_hash_common_crypto_h__
+#define INCLUDE_hash_common_crypto_h__
 
-#include "hash/sha1.h"
+#include "hash/sha.h"
 
-#include <mbedtls/sha1.h>
+#include <CommonCrypto/CommonDigest.h>
 
 struct git_hash_sha1_ctx {
-    mbedtls_sha1_context c;
+	CC_SHA1_CTX c;
 };
 
-#endif /* INCLUDE_hash_sha1_mbedtls_h__ */
+#endif
