+++ src/util/hash/openssl.c
