+++ src/util/hash/sha1dc/sha1.c
