+++ src/util/hash/win32.c
