+++ src/util/hash/win32.h
@@ -5,10 +5,10 @@
  * a Linking Exception. For full terms see the included COPYING file.
  */
 
-#ifndef INCLUDE_hash_sha1_win32_h__
-#define INCLUDE_hash_sha1_win32_h__
+#ifndef INCLUDE_hash_win32_h__
+#define INCLUDE_hash_win32_h__
 
-#include "hash/sha1.h"
+#include "hash/sha.h"
 
 #include <wincrypt.h>
 #include <strsafe.h>
